Fix visual studio build failure
Cast frame_size to be int for now to fix build warnings.

+++ vpx_scale/generic/yv12config.c
@@ -198,7 +198,7 @@ int vp9_realloc_frame_buffer(YV12_BUFFER_CONFIG *ybf,
       if (!ybf->buffer_alloc)
         return -1;
 
-      ybf->buffer_alloc_sz = frame_size;
+      ybf->buffer_alloc_sz = (int)frame_size;
 
       // This memset is needed for fixing valgrind error from C loop filter
       // due to access uninitialized memory in frame border. It could be
@@ -227,7 +227,7 @@ int vp9_realloc_frame_buffer(YV12_BUFFER_CONFIG *ybf,
     ybf->uv_stride = uv_stride;
 
     ybf->border = border;
-    ybf->frame_size = frame_size;
+    ybf->frame_size = (int)frame_size;
 
     ybf->y_buffer = ybf->buffer_alloc + (border * y_stride) + border;
     ybf->u_buffer = ybf->buffer_alloc + yplane_size +
